// src/redux/enums/ClienteEnum.ts
export enum ClienteEnum {
   POPULATE_CLIENTE_START = 'POPULATE_CLIENTE_START',
   POPULATE_CLIENTE_SUCCESS = 'POPULATE_CLIENTE_SUCCESS',
   POPULATE_CLIENTE_ERROR = 'POPULATE_CLIENTE_ERROR',

   ADD_CLIENTE_START = 'ADD_CLIENTE_START',
   ADD_CLIENTE_SUCCESS = 'ADD_CLIENTE_SUCCESS',
   ADD_CLIENTE_ERROR = 'ADD_CLIENTE_ERROR',

   DELETE_CLIENTE_START = 'DELETE_CLIENTE_START',
   DELETE_CLIENTE_SUCCESS = 'DELETE_CLIENTE_SUCCESS',
   DELETE_CLIENTE_ERROR = 'DELETE_CLIENTE_ERROR',

   UPDATE_CLIENTE_START = 'UPDATE_CLIENTE_START',
   UPDATE_CLIENTE_SUCCESS = 'UPDATE_CLIENTE_SUCCESS',
   UPDATE_CLIENTE_ERROR = 'UPDATE_CLIENTE_ERROR'
}